feat: Refactor image service imports and restructure service files for better organization
