+++ backend/src/services/image/caption-img-url.service.ts
