+++ backend/src/services/image/generate-img-prompt.service.ts
